requirements.txt
ai21==1.0.7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==2.4.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorcet==3.0.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
emoji==2.4.0
evaluate==0.4.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.1
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
holoviews==1.14.9
huggingface-hub==0.15.1
hvplot==0.7.3
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
openai==0.27.7
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandas-dq==1.28
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyamg==5.0.0
pyarrow==12.0.0
pycparser==2.21
pyct==0.5.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pyviz-comms==2.3.0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
requests==2.30.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
terminado==0.17.1
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.1
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
widgetsnbextension==4.0.7
wordcloud==1.9.2
xgboost==1.7.5
xlrd==2.0.1
xxhash==3.2.0
yarl==1.9.2
